requirements.txt
fastapi
uvicorn[standard]
yt-dlp
httpx
python-dateutil
aiofiles
python-memcached
